registry: Allow registry authentication for docker login
docker login happens with no scopes provided (unlike docker pull and
push, where there's always a scope). In this login case, provide
authentication for an empty scope. This means any username/password
is allowed for login, but nonetheless this does not weaken access, as
on pull and push the request will be rechecked whether it can access
the specific images.

This allows docker login to work again, that stopped working in
v0.17.4, that is #85

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@balena.io>