Add is_invalidated boolean term to the release resource
We currently need the flag to mark host OS releases as invalid
so they can be filtered out from the list of available releases.
The future plan is for the supervisor to automatically rollback
invalidated releases, but there are no dedicated resources to
work on that just yet.

Change-type: minor
Signed-off-by: Stevche Radevski <stevche@balena.io>