Add is_archived term to the application resource
The original purpose of the flag is to be able to
discontinue host applications, but the same concept
can be used by users to mark applications that are not
supported/developed anymore. This was discussed at a
product call and we decided to go with is_archived
based on Github's terminology as the meaning of both
are close.

Change-type: minor
Signed-off-by: Stevche Radevski <stevche@balena.io>