registry: Only resolve releases with one service
For the intial release we are not supporting applications with multiple
services as the intended use is blocks.

These changes will reject requests with a value for service following
the commit|semver.

It will also reject images at the DB query if the release has multiple
release images.

Revert this commit to restore support for multiple services.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>