Prevent unauthenticated requests from proceeding in token auth
Authentication failure due to "no credentials provided" and due to "invalid credentials provided" are currently treated in exactly the same way: we just let the request proceed and rely on (no) permissions to limit access to resources. This is problematic (it can result in redundant work being done in the API) and misleading ("I've provided creds but I get back no results").

This change separates the latter case and responds immediately with 401, as it should.

This is however a workaround and ideally it would be Pine that makes this distinction and throws an exception appropriately but this would be a breaking change to existing code.

Chat thread: https://jel.ly.fish/thread-79f9d489-8af5-42dd-b45f-84cef2d379a9

Change-type: patch