Fix env-var schema for backwards compatibility
PR #1199 relaxed the allowed character set for configuration variables,
but it also inadvertently relaxed the `/config/vars` JSON schema
exported by the API. This would cause that the UI allows the extended
character set for env vars, but the API would ultimately reject it,
creating an inconsistency. This PR adds a `configVarInvalidRegex`
variable to the schema that should be adopted by the UI to render
configuration variables.

Change-type: patch