Update @balena/pinejs to enable ISODatestring as internal pine API interface for date and date time fields.
Fix usage of `DateString` to actually as DateString and not as a `Date`

This is a major change as the exported internal API reference from `open-balena-api` when used as module breaks the type interface.
The model types have always been generated as `string` but actually returned as `Date`.

Change-type: major
Signed-off-by: fisehara <harald@balena.io>