Allow registering user with `null` email
user (Auth) email is a nullable property. However, instead of using the actual null current
API behaviour forces it to be an empty string. This is problematic as ideally the email field
would have a unique constraint to it (and NULL value is considered distinct by default,
different than empty string)

Change-type: minor