Target hostapp hook: Only patch should_be_operated_by__release if null or older release
With Supervisor managed HUP, we'll soon start relying on should_be_operated_by__release
to determine the target hostapp release. The Supervisor reports os_version and os_variant
on device provision, and the current target-hostapp hook will set the should_be_operated_by field
when receiving that initial state report. However, SV manage HUP means this field could be set
prior to device provision, so we don't want to overwrite it if it's not null.

If the should_be_operated_by field has a release that's older than the reported OS release
though, we should update the field to the reported OS release.

Change-type: patch
Signed-off-by: Christina Ying Wang <christina@balena.io>